+++ openvino_xai/explainer/explanation.py
@@ -4,12 +4,13 @@
 import os
 from enum import Enum
 from pathlib import Path
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import cv2
 import matplotlib.pyplot as plt
 import numpy as np
 
+from openvino_xai.common.parameters import Task
 from openvino_xai.common.utils import logger
 from openvino_xai.explainer.utils import (
     convert_targets_to_numpy,
@@ -36,6 +37,7 @@ class Explanation:
         saliency_map: np.ndarray | Dict[int | str, np.ndarray],
         targets: np.ndarray | List[int | str] | int | str,
         label_names: List[str] | None = None,
+        metadata: Dict[Task, Any] | None = None,
     ):
         targets = convert_targets_to_numpy(targets)
 
@@ -58,6 +60,7 @@ class Explanation:
             self._saliency_map = self._select_target_saliency_maps(targets, label_names)
 
         self.label_names = label_names
+        self.metadata = metadata
 
     @property
     def saliency_map(self) -> Dict[int | str, np.ndarray]: