+++ openvino_xai/methods/__init__.py
@@ -3,7 +3,8 @@
 """
 XAI algorithms.
 """
-from openvino_xai.methods.black_box.aise import AISE
+from openvino_xai.methods.black_box.aise.classification import AISEClassification
+from openvino_xai.methods.black_box.aise.detection import AISEDetection
 from openvino_xai.methods.black_box.rise import RISE
 from openvino_xai.methods.white_box.activation_map import ActivationMap
 from openvino_xai.methods.white_box.base import WhiteBoxMethod
@@ -24,5 +25,6 @@ __all__ = [
     "ViTReciproCAM",
     "DetClassProbabilityMap",
     "RISE",
-    "AISE",
+    "AISEClassification",
+    "AISEDetection",
 ]