+++ openvino_xai/metrics/__init__.py
@@ -3,3 +3,9 @@
 """
 Metrics in OpenVINO-XAI to check the quality of saliency maps.
 """
+
+from openvino_xai.metrics.adcc import ADCC
+from openvino_xai.metrics.insertion_deletion_auc import InsertionDeletionAUC
+from openvino_xai.metrics.pointing_game import PointingGame
+
+__all__ = ["ADCC", "InsertionDeletionAUC", "PointingGame"]