+++ tests/intg/test_accuracy_metrics.py
@@ -5,6 +5,7 @@ import json
 from typing import Dict, List, Tuple
 
 import cv2
+import numpy as np
 import openvino as ov
 import pytest
 
@@ -16,9 +17,8 @@ from openvino_xai.explainer.utils import (
     get_postprocess_fn,
     get_preprocess_fn,
 )
-from openvino_xai.metrics.insertion_deletion_auc import InsertionDeletionAUC
-from openvino_xai.metrics.pointing_game import PointingGame
-from tests.unit.explanation.test_explanation_utils import VOC_NAMES
+from openvino_xai.metrics import ADCC, InsertionDeletionAUC, PointingGame
+from tests.unit.explainer.test_explanation_utils import VOC_NAMES
 
 MODEL_NAME = "mlc_mobilenetv3_large_voc"
 IMAGE_PATH = "tests/assets/cheetah_person.jpg"
@@ -54,18 +54,9 @@ def load_gt_bboxes(json_coco_path: str) -> List[Dict[str, List[Tuple[int, int, i
     return list(result.values())
 
 
-class TestDummyRegression:
+class TestAccuracyMetrics:
     image = cv2.imread(IMAGE_PATH)
     gt_bboxes = load_gt_bboxes(COCO_ANN_PATH)
-    pointing_game = PointingGame()
-
-    preprocess_fn = get_preprocess_fn(
-        change_channel_order=True,
-        input_size=(224, 224),
-        hwc_to_chw=True,
-    )
-
-    postprocess_fn = get_postprocess_fn(activation=ActivationType.SIGMOID)
 
     @pytest.fixture(autouse=True)
     def setup(self, fxt_data_root):
@@ -73,38 +64,56 @@ class TestDummyRegression:
         retrieve_otx_model(data_dir, MODEL_NAME)
         model_path = data_dir / "otx_models" / (MODEL_NAME + ".xml")
         model = ov.Core().read_model(model_path)
-        compiled_model = ov.Core().compile_model(model, "CPU")
-        self.auc = InsertionDeletionAUC(compiled_model, self.preprocess_fn, self.postprocess_fn)
+
+        preprocess_fn = get_preprocess_fn(
+            change_channel_order=True,
+            input_size=(224, 224),
+            hwc_to_chw=True,
+        )
+        postprocess_fn = get_postprocess_fn(activation=ActivationType.SIGMOID)
 
         self.explainer = Explainer(
             model=model,
             task=Task.CLASSIFICATION,
-            preprocess_fn=self.preprocess_fn,
+            preprocess_fn=preprocess_fn,
             explain_mode=ExplainMode.WHITEBOX,
         )
+
+        self.pointing_game = PointingGame()
+        self.auc = InsertionDeletionAUC(model, preprocess_fn, postprocess_fn)
+        self.adcc = ADCC(model, preprocess_fn, postprocess_fn, self.explainer)
 
     def test_explainer_image(self):
         explanation = self.explainer(self.image, targets=["person"], label_names=VOC_NAMES, colormap=False)
         assert len(explanation.saliency_map) == 1
+
         pointing_game_score = self.pointing_game.evaluate([explanation], self.gt_bboxes)["pointing_game"]
         assert pointing_game_score == 1.0
 
-        explanation = self.explainer(self.image, targets=["person"], label_names=VOC_NAMES, colormap=False)
-        assert len(explanation.saliency_map) == 1
         auc_score = self.auc.evaluate([explanation], [self.image], steps=10).values()
         insertion_auc_score, deletion_auc_score, delta_auc_score = auc_score
-        assert insertion_auc_score >= 0.9
-        assert deletion_auc_score >= 0.2
-        assert delta_auc_score >= 0.7
+        assert np.abs(insertion_auc_score - 0.94) <= 0.01
+        assert np.abs(deletion_auc_score - 0.23) <= 0.01
+        assert np.abs(delta_auc_score - 0.71) <= 0.01
 
-        # Two classes for saliency maps
+        adcc_score = self.adcc.evaluate([explanation], [self.image])["adcc"]
+        assert np.abs(adcc_score - 0.96) <= 0.01
+
+    def test_explainer_image_2_classes(self):
         explanation = self.explainer(self.image, targets=["person", "cat"], label_names=VOC_NAMES, colormap=False)
         assert len(explanation.saliency_map) == 2
+
+        pointing_game_score = self.pointing_game.evaluate([explanation], self.gt_bboxes)["pointing_game"]
+        assert pointing_game_score == 1.0
+
         auc_score = self.auc.evaluate([explanation], [self.image], steps=10).values()
         insertion_auc_score, deletion_auc_score, delta_auc_score = auc_score
-        assert insertion_auc_score >= 0.5
-        assert deletion_auc_score >= 0.1
-        assert delta_auc_score >= 0.35
+        assert np.abs(insertion_auc_score - 0.53) <= 0.01
+        assert np.abs(deletion_auc_score - 0.13) <= 0.01
+        assert np.abs(delta_auc_score - 0.39) <= 0.01
+
+        adcc_score = self.adcc.evaluate([explanation], [self.image])["adcc"]
+        assert np.abs(adcc_score - 0.55) <= 0.01
 
     def test_explainer_images(self):
         images = [self.image, self.image]
@@ -119,6 +128,9 @@ class TestDummyRegression:
 
         auc_score = self.auc.evaluate(explanations, images, steps=10).values()
         insertion_auc_score, deletion_auc_score, delta_auc_score = auc_score
-        assert insertion_auc_score >= 0.9
-        assert deletion_auc_score >= 0.2
-        assert delta_auc_score >= 0.7
+        assert np.abs(insertion_auc_score - 0.94) <= 0.01
+        assert np.abs(deletion_auc_score - 0.23) <= 0.01
+        assert np.abs(delta_auc_score - 0.71) <= 0.01
+
+        adcc_score = self.adcc.evaluate(explanations, images)["adcc"]
+        assert np.abs(adcc_score - 0.96) <= 0.01